+++ helpers/account_helper.py
@@ -15,7 +15,6 @@ class AccountHelper:
             'email': email,
             'password': password,
         }
-
         response = self.dm_account_api.account_api.post_v1_account(json_data=json_data)
         assert response.status_code == 201, f"Пользователь не был создан {response.json()}"
         response = self.mailhog.mailhog_api.get_api_v2_messages()
@@ -32,10 +31,39 @@ class AccountHelper:
             'password': password,
             'rememberMe': remember_me,
         }
-
         response = self.dm_account_api.login_api.post_v1_account_login(json_data=json_data)
         assert response.status_code == 200, "Пользователь не смог авторизоваться"
         return response
+
+    def change_email(self, login: str, password: str, email: str):
+        email = f'string_{email}'
+        json_data = {
+            'login': login,
+            'email': email,
+            'password': password,
+        }
+        response = self.dm_account_api.account_api.put_v1_account_email(json_data)
+        assert response.status_code == 200, "Почтовый ящик не был сменён"
+        response = self.mailhog.mailhog_api.get_api_v2_messages()
+        assert response.status_code == 200, "Письма не были получены"
+        token = self.get_new_activation_token_by_login(login, response, email)
+        assert token is not None, f"Токен для пользователя {login} не был получен"
+        response = self.dm_account_api.account_api.put_v1_account_token(token=token)
+        assert response.status_code == 200, "Пользователь не был активирован"
+        return response
+
+    def get_auth_account_token(self, login: str, password: str, email: str):
+        json_data = {
+            'login': login,
+            'email': email,
+            'password': password,
+        }
+        response = self.dm_account_api.account_api.post_v1_account(json_data=json_data)
+        assert response.status_code == 201, f"Пользователь не был создан {response.json()}"
+        response = self.mailhog.mailhog_api.get_api_v2_messages()
+        assert response.status_code == 200, "Письма не были получены"
+        token = self.get_activation_token_by_login(login=login, response=response)
+        assert token is not None, f"Токен для пользователя {login} не был получен"
 
     @staticmethod
     def get_activation_token_by_login(
@@ -49,4 +77,21 @@ class AccountHelper:
 
             if user_login == login:
                 token = user_data['ConfirmationLinkUrl'].split('/')[-1]
+        return token
+
+    @staticmethod
+    def get_new_activation_token_by_login(
+            login,
+            response,
+            email
+    ):
+        token = None
+        for item in response.json()['items']:
+            user_data = loads(item['Content']['Body'])
+            user_login = user_data['Login']
+            user_new_email = item['Content']['Headers']['To'][0]
+
+            if user_login == login and user_new_email == email:
+                token = user_data['ConfirmationLinkUrl'].split('/')[-1]
+
         return token
