+++ tests/functional/post_v1_account_login/test_post_v1_account_login.py
@@ -1,13 +1,13 @@
-from json import loads
-import random
+from helpers.account_helper import AccountHelper
 
-from dm_api_account.apis.account_api import AccountApi
-from dm_api_account.apis.login_api import LoginApi
-from api_mailhog.apis.mailhog_api import MailhogApi
-
-import structlog
 from restclient.configuration import Configuration as MailhogConfiguration
 from restclient.configuration import Configuration as DmApiConfiguration
+
+from services.dm_api_account import DMApiAccount
+from services.api_mailhog import MailHogApi
+
+import random
+import structlog
 
 structlog.configure(
     processors=[
@@ -21,60 +21,17 @@ structlog.configure(
 
 
 def test_post_v1_account_login():
-    # Регистрация пользователя
     mailhog_configuration = MailhogConfiguration(host='http://5.63.153.31:5025')
     dm_api_configuration = DmApiConfiguration(host='http://5.63.153.31:5051', disable_log=False)
 
-    account_api = AccountApi(configuration=dm_api_configuration)
-    login_api = LoginApi(configuration=dm_api_configuration)
-    mailhog_api = MailhogApi(configuration=mailhog_configuration)
+    account = DMApiAccount(configuration=dm_api_configuration)
+    mailhog = MailHogApi(configuration=mailhog_configuration)
+
+    account_helper = AccountHelper(dm_account_api=account, mailhog=mailhog)
 
     login = f'morugova_test.{random.random()}'
     password = '123456789'
     email = f'{login}@mail.ru'
-    json_data = {
-        'login': login,
-        'email': email,
-        'password': password,
-    }
 
-    response = account_api.post_v1_account(json_data=json_data)
-    assert response.status_code == 201, f"Пользователь не был создан {response.json()}"
-
-    # Получить письма из почтового сервера
-
-    response = mailhog_api.get_api_v2_messages()
-    assert response.status_code == 200, "Письма не были получены"
-
-    # Получить активационный токен
-    token = get_activation_token_by_login(login, response)
-
-    assert token is not None, f"Токен для пользователя {login} не был получен"
-
-    # Активация пользователя
-    response = account_api.put_v1_account_token(token=token)
-    assert response.status_code == 200, "Пользователь не был активирован"
-
-    # # Авторизоваться
-    json_data = {
-        'login': login,
-        'password': password,
-        'rememberMe': True,
-    }
-
-    response = login_api.post_v1_account_login(json_data=json_data)
-    assert response.status_code == 200, "Пользователь не смог авторизоваться"
-
-
-def get_activation_token_by_login(
-        login,
-        response
-):
-    token = None
-    for item in response.json()['items']:
-        user_data = loads(item['Content']['Body'])
-        user_login = user_data['Login']
-
-        if user_login == login:
-            token = user_data['ConfirmationLinkUrl'].split('/')[-1]
-    return token
+    account_helper.register_new_user(login=login, password=password, email=email)
+    account_helper.user_login(login=login, password=password)
